Rename auth.ts to jwtAuth.ts and update imports
- Renamed `auth.ts` to `jwtAuth.ts` for better clarity and alignment with its purpose.
- Updated all import references from `auth` to `jwtAuth` across the codebase to prevent any bugs.
- Ensures consistent naming for authentication utilities using JWT.
